requirements.txt
asgiref==3.6.0
autopep8==2.0.2
colorama==0.4.6
coverage==7.2.3
Django==4.2
django-ckeditor==6.5.1
django-debug-toolbar==4.0.0
django-js-asset==2.0.0
factory-boy==3.2.1
Faker==18.4.0
gunicorn==20.1.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
